refactor(intents): seperated model/source specific intents & actions into seperate modules: - fromDOM & fromSocketIO modules added accordingly to node & feed folders - moved old unused actions to "attic" - various related changes and tweaks
