+++ src/client/core/nodes/fromSocketIO.js
@@ -1,0 +1,12 @@
+import {toArray} from '../../utils/utils'
+
+export default function actions(socketIO){
+  const setNodes$ =
+    socketIO.get("initialData")
+    .map(e=>JSON.parse(e))
+    .map(data=>({data}))
+
+  return {
+    setNodes$
+  }
+}
